Create Array.prototype.last
# Intuition
Arrays in JavaScript can be extended using the `prototype` property. To solve this problem, we want every array to have a `.last()` method that returns the last element. If the array is empty, it should return `-1`. This behavior is similar to checking the last element of an array in C, but more flexible.

# Approach
We define a method `last()` on `Array.prototype`, which allows all array instances to access it. Inside this method:
- If `this.length` is `0`, we return `-1`.
- Otherwise, we return `this[this.length - 1]`, which is the last element in the array.

This approach ensures the method works for any array, including ones parsed from JSON.

# Complexity
- Time complexity:  
  $$O(1)$$  
  Accessing the last element or checking the length is done in constant time.

- Space complexity:  
  $$O(1)$$  
  We don't use any additional memory beyond the function definition.